requirements.txt
Django
djangorestframework
requests
mysqlclient